# user-service/src/main/java/com/Finzo/user_service/dto/UserResponse.java
package com.Finzo.user_service.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.RequiredArgsConstructor;

@Data
@AllArgsConstructor
@RequiredArgsConstructor
public class UserResponse {

    String name;
    String email;
    Long id;
    String password;
}